wowless: unify the two containment-relation implementations
tools/xmlcontainment.lua's supertypesOf/childrenOf and tools/prep.lua's
xmlflat builder independently reimplemented the exact same
extends-chain climb (respecting sealed) to compute a tag's supertypes
and accepted content -- two implementations of the same relation, with
nothing to keep them in agreement if one changed without the other.

Moves supertypesOf/childrenOf into a new wowless/xmlutil.lua, required
by both: tools/xmlcontainment.lua's chains/legalChildren/roots build on
it directly, and tools/prep.lua's xmlflat builder now calls it instead
of climbing the chain a second, separate way. xmlflat is what
wowless/modules/xml.lua's runtime parser actually consumes, so this
also means the runtime's real containment behavior and the build-time
tools that reason about the same schema are now provably the same
computation, not just two implementations that happen to agree today.

childrenOf gained a second return value, `text`, so prep.lua doesn't
need its own separate climb to determine a tag's flattened text-vs-tags
content -- existing callers that only want the children set are
unaffected, since Lua discards extra return values in a single-value
context.

Verified the two are genuinely the same implementation, not just
similarly-shaped code: temporarily broke childrenOf's key computation
and confirmed it fails all three of spec/tools/xmlcontainment_spec.lua,
spec/wowless/modules/xml_containment_spec.lua, and -- critically --
the pre-existing spec/wowless/modules/xml_spec.lua, which exercises
wowless/modules/xml.lua's actual runtime parsing behavior. Reverted
before committing.
